asyncio fixtures and tests
